Move FormFactory namespace to Form\Factory
